Add support for built-in templates; fix #419
